% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-NETseqData.R
\name{.process_colData_scores}
\alias{.process_colData_scores}
\title{Optionally Apply masks to a scores GPos and/or zero fill the gaps}
\usage{
.process_colData_scores(u, apply_mask, zero_fill, range_scope = NULL)
}
\description{
Optionally Apply masks to a scores GPos and/or zero fill the gaps
}
