% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPosExperiment-package.R
\docType{package}
\name{GPosExperiment-package}
\alias{GPosExperiment}
\alias{GPosExperiment-package}
\title{GPosExperiment: Segmentation Analysis Framework for Genomic Sequences}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Robert Shear \email{rshear@gmail.com} (\href{https://orcid.org/000-0003-2680-267X}{ORCID})

}
\keyword{internal}
