% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HRseqData-class}
\alias{HRseqData-class}
\title{Contains sample-level assay date}
\description{
TBD
}
\section{Slots}{

\describe{
\item{\code{sampleId}}{}

\item{\code{scores}}{}

\item{\code{mask}}{}

\item{\code{seqinfo}}{}
}}

