% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-HRseqData.R
\name{HRseqData}
\alias{HRseqData}
\title{HRseqData - Construct HRseqData}
\usage{
HRseqData(
  scores = GRanges(),
  mask = GRanges(),
  sampleId = character(),
  seqinfo = NULL,
  ...
)
}
\arguments{
\item{scores}{}

\item{mask}{}

\item{sampleId}{}

\item{...}{}
}
\description{
TBD
}
