% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRseqDataFromBAM.R
\name{HRseqDataFromBAM,character-method}
\alias{HRseqDataFromBAM,character-method}
\title{HRseqDataFromBAM}
\usage{
\S4method{HRseqDataFromBAM}{character}(sampleId, bam_file, gene_list, seqinfo)
}
\arguments{
\item{sampleId}{String}

\item{bam_file}{Location of BAM file}

\item{gene_list}{Filename of gff3 file containing regions of interest}

\item{seqinfo}{a SeqInfo object for the appropriate genome}
}
\value{
A HRseqData object
}
\description{
Create a list of HRseqData objects from a list of BAM files
}
\details{
TODO - more description
}
