% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HRseqExperiment-class}
\alias{HRseqExperiment-class}
\alias{.HRseqExperiment}
\title{An S4 subclass of RangedSummarizedExperiment}
\description{
TBD
}
\section{Slots}{

\describe{
\item{\code{seqinfo}}{}
}}

