% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRseqExperiment-package.R
\docType{package}
\name{HRseqExperiment-package}
\alias{HRseqExperiment}
\alias{HRseqExperiment-package}
\title{HRseqExperiment:}
\description{
The HRseqExperiment class is an extension of RangedSummarizedExperiment equipped to process data from high-throughput sequencing assays where measurements at nucleotide level resolution is required.
}
\author{
\strong{Maintainer}: Robert Shear \email{rshear@gmail.com} (\href{https://orcid.org/000-0003-2680-267X}{ORCID})

}
\keyword{internal}
