% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-NETseqData.R
\name{NETseqData}
\alias{NETseqData}
\title{NETseqData Constructor}
\usage{
NETseqData(
  scores = GPos(stitch = FALSE),
  segments = GRanges(),
  sampleId = character(),
  seqinfo = NULL,
  ...
)
}
\arguments{
\item{scores}{}

\item{segments}{}

\item{sampleId}{}

\item{...}{}
}
\description{
NETseqData Constructor
}
