% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NETseqDataFromBAM.R
\name{NETseqDataFromBAM,character-method}
\alias{NETseqDataFromBAM,character-method}
\title{NETseqDataFromBAM}
\usage{
\S4method{NETseqDataFromBAM}{character}(sampleId, bam_file, gene_list, seqinfo)
}
\description{
NETseqDataFromBAM
}
