% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-GPosExperiment.R
\name{scores,GPosExperiment-method}
\alias{scores,GPosExperiment-method}
\title{Get  GPosExperiment scores vector}
\usage{
\S4method{scores}{GPosExperiment}(x, apply_mask = TRUE, zero_fill = TRUE)
}
\arguments{
\item{zero_fill}{}
}
\description{
TODO Long Title
}
\details{
TODO Narraitve
}
