% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-HRseqData.R
\name{scores,HRseqData-method}
\alias{scores,HRseqData-method}
\title{scores -   HRseqData scores vector}
\usage{
\S4method{scores}{HRseqData}(x, apply_mask = FALSE, zero_fill = FALSE, range_scope = NULL)
}
\arguments{
\item{range_scope}{}
}
\description{
TODO Long Title
}
\details{
TODO Narraitve
}
