% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-NETseqData.R
\name{scores,NETseqData-method}
\alias{scores,NETseqData-method}
\title{Get  NETseqData scores vector}
\usage{
\S4method{scores}{NETseqData}(x, apply_mask = FALSE, zero_fill = FALSE, range_scope = NULL)
}
\arguments{
\item{range_scope}{}
}
\description{
TODO Long Title
}
\details{
TODO Narraitve
}
