% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-GPosExperiment.R
\name{segments,GPosExperiment-method}
\alias{segments,GPosExperiment-method}
\title{segments}
\usage{
\S4method{segments}{GPosExperiment}(x, reverse.negative = FALSE, ...)
}
\description{
TODO: describe
}
