% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-NETseqData.R
\name{NETseqData}
\alias{NETseqData}
\title{NETseqData - Construct list of NETseqData objects}
\usage{
NETseqData(
  scores = GRanges(),
  mask = GRanges(),
  sampleId = character(),
  seqinfo = NULL,
  ...
)
}
\arguments{
\item{scores}{}

\item{mask}{}

\item{sampleId}{}

\item{...}{}
}
\description{
TBD
}
