% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NETseqDataFromBAM.R
\name{NETseqDataFromBAM,character-method}
\alias{NETseqDataFromBAM,character-method}
\title{NETseqDataFromBAM}
\usage{
\S4method{NETseqDataFromBAM}{character}(sampleId, bam_file, gene_list, seqinfo)
}
\arguments{
\item{sampleId}{String}

\item{bam_file}{Location of BAM file}

\item{gene_list}{Filename of gff3 file containing regions of interest}

\item{seqinfo}{a SeqInfo object for the appropriate genome}
}
\value{
A NETseqData object
}
\description{
Create a list of NETseqData objects from a list of BAM files
}
\details{
TODO - more description
}
