% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-HRseqExperiment.R
\name{scores,HRseqExperiment-method}
\alias{scores,HRseqExperiment-method}
\title{Get  HRseqExperiment scores vector}
\usage{
\S4method{scores}{HRseqExperiment}(x, apply_mask = TRUE, zero_fill = TRUE)
}
\arguments{
\item{zero_fill}{}
}
\description{
TODO Long Title
}
\details{
TODO Narraitve
}
